requirements.txt
ursina
noise
Pillow